+++ src/logger/logger_module.py
@@ -1,8 +1,10 @@
 import os
-import logging 
+import logging
 
 
-def create_logger(path_logfile : str, name_logfile : str, mode='w') -> logging.FileHandler:
+def create_logger(
+    path_logfile: str, name_logfile: str, mode="w"
+) -> logging.FileHandler:
     """Function to create logger for the bot
 
     Args:
@@ -12,8 +14,9 @@ def create_logger(path_logfile : str, name_logfile : str, mode='w') -> logging.F
     Returns:
         logging.FileHandler: Filehandler for logging purposes
     """
-    
+
     # Build logger output from the bot
     os.makedirs(name=path_logfile, exist_ok=True)
-    return logging.FileHandler(filename=os.path.join(path_logfile, name_logfile), mode=mode)
-    
+    return logging.FileHandler(
+        filename=os.path.join(path_logfile, name_logfile), mode=mode
+    )