+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## 0.5.0 (2026-05-23)
+
+### Refactor
+
+- **main**: logger on INFO level
+
+## 0.4.0 (2026-05-23)
+
+### Feat
+
+- **src**: Replaced prefix commands with slash commands
+- **src**: added admin only commands to handle database
+- **events**: added tasks to retrieve log & data to admin server
+- **src**: added mention messages
+- **src**: updated command to get specific quests label
+- **src**: use guild display name
+- **src**: added command to read all quests in server
+- **src**: added help command + aliases
+- **src**: labels + help command
+- **label**: added labels dict
+
+### Fix
+
+- **database**: replace status code on remove
+
 ## 0.3.0 (2026-05-20)
 
 ### BREAKING CHANGE